python/requirements.txt
trafilatura
pika
prometheus-client
requests
warcio
minio
langdetect
pyyaml
pycld2
pandas
pyarrow